# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LAPES - API",
    "description": "API for supervisory control and data acquisition",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
    "/sessions/password": {
      "post": {
        "summary": "Authenticate with e-mail & password",
        "tags": ["Auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string", "minLength": 6 }
                },
                "required": ["username", "password"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Login realizado, token JWT enviado",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Login realizado, token JWT enviado",
                  "type": "object",
                  "properties": { "token": { "type": "string" } },
                  "required": ["token"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Credenciais inválidas",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/meter/{id}": {
      "patch": {
        "summary": "Update meter enabled status",
        "tags": ["Meters"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "enabled": { "type": "boolean" } },
                "required": ["enabled"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": { "type": "number" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Estado do medidor atualizado" },
          "400": { "description": "Falha ao atualizar estado do medidor" },
          "401": {
            "description": "Não autorizado, necessita de login",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Não autorizado, necessita de login",
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/meters/{id}": {
      "put": {
        "summary": "Update an existing meter",
        "tags": ["Meters"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "ip": { "type": "string", "minLength": 1 },
                  "description": { "type": "string" }
                },
                "required": ["name", "ip"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": { "type": "number" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Sucesso",
                  "type": "object",
                  "properties": {},
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado, necessita de login",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Não autorizado, necessita de login",
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Medidor não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Medidor não encontrado",
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a meter",
        "tags": ["Meters"],
        "parameters": [
          {
            "schema": { "type": "number" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Sucesso",
                  "type": "object",
                  "properties": {},
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado, necessita de login",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Não autorizado, necessita de login",
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Medidor não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Medidor não encontrado",
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}